Restructure pipeline packages and fix minor issues
Consolidate 6 packages down to 3: merge DownloadService into FeedService,
group CLI wrappers (AudioConverter, Transcriber) under external/, and flatten
AppConfig and Utils to the root package. Also replace SimpleDateFormat with
thread-safe DateTimeFormatter and remove stale transcript_with_timing.tsv
from README output structure.
