+++ pipeline/src/main/kotlin/com/rsstowhisper/AppConfig.kt
@@ -1,4 +1,4 @@
-package com.rsstowhisper.config
+package com.rsstowhisper
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties
 import com.fasterxml.jackson.annotation.JsonProperty
