+++ pipeline/src/main/kotlin/com/rsstowhisper/external/TranscriptWriter.kt
@@ -1,4 +1,4 @@
-package com.rsstowhisper.transcription
+package com.rsstowhisper.external
 
 import java.nio.file.Files
 import java.nio.file.Path
